fix: extract auth options and repair build
