+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,42 +1,5 @@
-import NextAuth, { NextAuthOptions } from "next-auth";
-import Credentials from "next-auth/providers/credentials";
-import { PrismaAdapter } from "@next-auth/prisma-adapter";
-import prisma from "@/lib/prisma";
-
-export const authOptions: NextAuthOptions = {
-  adapter: PrismaAdapter(prisma),
-  providers: [
-    Credentials({
-      name: "Credentials",
-      credentials: {
-        email: { label: "Email", type: "email" },
-        password: { label: "Password", type: "password" }
-      },
-      async authorize(credentials) {
-        if (!credentials?.email) return null;
-        // Dev-only: allow login if the user exists by email
-        const user = await prisma.user.findUnique({ where: { email: credentials.email } });
-        return user ?? null;
-      }
-    })
-  ],
-  session: { strategy: "jwt" },
-  callbacks: {
-    async jwt({ token, user }) {
-      if (user && "role" in user) {
-        token.role = (user as { role?: string }).role;
-      }
-      return token;
-    },
-    async session({ session, token }) {
-      const user = (session.user ?? {}) as typeof session.user & { role?: string };
-      session.user = { ...user, role: (token as { role?: string }).role };
-      return session;
-    }
-  },
-  // Use a constant fallback secret in development to avoid JWT decryption errors
-  secret: process.env.NEXTAUTH_SECRET ?? "insecure-dev-secret"
-};
+import NextAuth from "next-auth";
+import { authOptions } from "@/lib/auth";
 
 const handler = NextAuth(authOptions);
 export { handler as GET, handler as POST };
