+++ src/app/layout.tsx
@@ -8,10 +8,10 @@ export default function RootLayout({ children }: { children: React.ReactNode }) 
   return (
     <html lang="en">
       <body className="min-h-dvh antialiased">
-        {/* @ts-expect-error Server Component */}
-        <AppNav />
-        <main className="container py-8"><div className="mx-auto max-w-6xl space-y-8">{children}</div></main>
-        <Providers>{/* client providers here */}</Providers>
+        <Providers>
+          <AppNav />
+          <main className="container py-8"><div className="mx-auto max-w-6xl space-y-8">{children}</div></main>
+        </Providers>
       </body>
     </html>
   );
