+++ src/lib/auth.ts
@@ -1,7 +1,41 @@
-import { getServerSession } from "next-auth";
-import { authOptions } from "@/app/api/auth/[...nextauth]/route";
+import { NextAuthOptions, getServerSession } from "next-auth";
+import Credentials from "next-auth/providers/credentials";
+import { PrismaAdapter } from "@next-auth/prisma-adapter";
+import prisma from "@/lib/prisma";
+
+export const authOptions: NextAuthOptions = {
+  adapter: PrismaAdapter(prisma),
+  providers: [
+    Credentials({
+      name: "Credentials",
+      credentials: {
+        email: { label: "Email", type: "email" },
+        password: { label: "Password", type: "password" }
+      },
+      async authorize(credentials) {
+        if (!credentials?.email) return null;
+        const user = await prisma.user.findUnique({ where: { email: credentials.email } });
+        return user ?? null;
+      }
+    })
+  ],
+  session: { strategy: "jwt" },
+  callbacks: {
+    async jwt({ token, user }) {
+      if (user && "role" in user) {
+        token.role = (user as { role?: string }).role;
+      }
+      return token;
+    },
+    async session({ session, token }) {
+      const user = (session.user ?? {}) as typeof session.user & { role?: string };
+      session.user = { ...user, role: (token as { role?: string }).role } as typeof session.user;
+      return session;
+    }
+  },
+  secret: process.env.NEXTAUTH_SECRET ?? "insecure-dev-secret"
+};
 
 export function getSession() {
   return getServerSession(authOptions);
 }
-
